fix: fonts, All fonts should com from next/font
