+++ src/app/globals.css
@@ -1,7 +1,6 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-@import '@fontsource/inter';
 
 /* :root {
   --foreground-rgb: 0, 0, 0;
